KL-72/refactor: change convert to converter
