+++ src/main/java/taco/klkl/domain/region/converter/RegionTypeConverter.java
@@ -1,11 +1,11 @@
-package taco.klkl.domain.region.convert;
+package taco.klkl.domain.region.converter;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import taco.klkl.domain.region.enums.RegionType;
 
 @Converter(autoApply = true)
-public class RegionTypeConvert implements AttributeConverter<RegionType, String> {
+public class RegionTypeConverter implements AttributeConverter<RegionType, String> {
 
 	@Override
 	public String convertToDatabaseColumn(RegionType regionType) {
